implement deleting archived recordings
