+++ src/main/resources/types.graphqls
@@ -81,7 +81,7 @@ type ArchivedRecording implements Recording {
     reportUrl: Url!
     downloadUrl: Url!
 
-    # delete: ArchivedRecording! # mutation
+    delete: ArchivedRecording! # mutation
 }
 
 interface Recording {
